Add all Assignment before/in/after in Lesson
